requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose
redis
stripe
httpx
